requirements.txt
# myPicasa 專案依賴套件
# 安裝方式: pip install -r requirements.txt

# GUI 框架
PyQt5>=5.15.0

# 圖片處理
Pillow>=9.0.0

# 影片處理
moviepy>=1.0.3

# 自然排序
natsort>=8.0.0
